backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
gunicorn==21.2.0
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
requests==2.31.0
